refactor(lib): move IDs to own file from "types"
