+++ lib/src/types.rs
@@ -1,6 +1,7 @@
 use std::sync::Arc;
 
 use crate::config::v2::tui::{keys::KeyBinding, theme::styles::ColorTermusic};
+use crate::ids::{IdConfigEditor, IdKey};
 use crate::invidious::{Instance, YoutubeVideo};
 use crate::podcast::{EpData, PodcastFeed, PodcastNoId};
 use crate::songtag::SongTag;
@@ -564,187 +565,6 @@ pub enum TFMsg {
     TableLyricOptionsBlurUp,
     TextareaLyricBlurDown,
     TextareaLyricBlurUp,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
-pub enum Id {
-    ConfigEditor(IdConfigEditor),
-    DBListCriteria,
-    DBListSearchResult,
-    DBListSearchTracks,
-    DeleteConfirmRadioPopup,
-    DeleteConfirmInputPopup,
-    DownloadSpinner,
-    Episode,
-    ErrorPopup,
-    GeneralSearchInput,
-    GeneralSearchTable,
-    GlobalListener,
-    HelpPopup,
-    Label,
-    Library,
-    Lyric,
-    MessagePopup,
-    Playlist,
-    Podcast,
-    PodcastAddPopup,
-    PodcastSearchTablePopup,
-    FeedDeleteConfirmRadioPopup,
-    FeedDeleteConfirmInputPopup,
-    Progress,
-    QuitPopup,
-    SavePlaylistPopup,
-    SavePlaylistLabel,
-    SavePlaylistConfirm,
-    TagEditor(IdTagEditor),
-    YoutubeSearchInputPopup,
-    YoutubeSearchTablePopup,
-    DatabaseAddConfirmPopup,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
-pub enum IdTagEditor {
-    CounterDelete,
-    LabelHint,
-    InputArtist,
-    InputTitle,
-    InputAlbum,
-    InputGenre,
-    SelectLyric,
-    TableLyricOptions,
-    TextareaLyric,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
-pub enum IdConfigEditor {
-    AlbumPhotoAlign,
-    CEThemeSelect,
-    ConfigSavePopup,
-    ExitConfirmation,
-    Footer,
-    Header,
-    Key(IdKey),
-    KillDamon,
-
-    LibraryBackground,
-    LibraryBorder,
-    LibraryForeground,
-    LibraryHighlight,
-    LibraryHighlightSymbol,
-    LibraryLabel,
-
-    LyricBackground,
-    LyricBorder,
-    LyricForeground,
-    LyricLabel,
-
-    MusicDir,
-    PlayerPort,
-    PlayerUseDiscord,
-    PlayerUseMpris,
-
-    PlaylistBackground,
-    PlaylistBorder,
-    PlaylistDisplaySymbol,
-    PlaylistForeground,
-    PlaylistHighlight,
-    PlaylistHighlightSymbol,
-    PlaylistLabel,
-    PlaylistRandomAlbum,
-    PlaylistRandomTrack,
-
-    CurrentlyPlayingTrackSymbol,
-
-    PodcastDir,
-    PodcastMaxRetries,
-    PodcastSimulDownload,
-
-    ProgressBackground,
-    ProgressBorder,
-    ProgressForeground,
-    ProgressLabel,
-
-    SaveLastPosition,
-    SeekStep,
-
-    ImportantPopupLabel,
-    ImportantPopupBackground,
-    ImportantPopupBorder,
-    ImportantPopupForeground,
-
-    FallbackBackground,
-    FallbackBorder,
-    FallbackForeground,
-    FallbackHighlight,
-    FallbackLabel,
-}
-
-#[derive(Debug, Eq, PartialEq, Clone, Copy, Hash)]
-pub enum IdKey {
-    DatabaseAddAll,
-    DatabaseAddSelected,
-    GlobalConfig,
-    GlobalDown,
-    GlobalGotoBottom,
-    GlobalGotoTop,
-    GlobalHelp,
-    GlobalLayoutTreeview,
-    GlobalLayoutDatabase,
-    GlobalLeft,
-    GlobalLyricAdjustForward,
-    GlobalLyricAdjustBackward,
-    GlobalLyricCycle,
-    GlobalPlayerToggleGapless,
-    GlobalPlayerTogglePause,
-    GlobalPlayerNext,
-    GlobalPlayerPrevious,
-    GlobalPlayerSeekForward,
-    GlobalPlayerSeekBackward,
-    GlobalPlayerSpeedUp,
-    GlobalPlayerSpeedDown,
-    GlobalQuit,
-    GlobalRight,
-    GlobalUp,
-    GlobalVolumeDown,
-    GlobalVolumeUp,
-    GlobalSavePlaylist,
-    LibraryDelete,
-    LibraryLoadDir,
-    LibraryPaste,
-    LibrarySearch,
-    LibrarySearchYoutube,
-    LibraryTagEditor,
-    LibraryYank,
-    PlaylistDelete,
-    PlaylistDeleteAll,
-    PlaylistShuffle,
-    PlaylistModeCycle,
-    PlaylistPlaySelected,
-    PlaylistSearch,
-    PlaylistSwapDown,
-    PlaylistSwapUp,
-    PlaylistAddRandomAlbum,
-    PlaylistAddRandomTracks,
-    LibrarySwitchRoot,
-    LibraryAddRoot,
-    LibraryRemoveRoot,
-    GlobalLayoutPodcast,
-    GlobalXywhMoveLeft,
-    GlobalXywhMoveRight,
-    GlobalXywhMoveUp,
-    GlobalXywhMoveDown,
-    GlobalXywhZoomIn,
-    GlobalXywhZoomOut,
-    GlobalXywhHide,
-    PodcastMarkPlayed,
-    PodcastMarkAllPlayed,
-    PodcastEpDownload,
-    PodcastEpDeleteFile,
-    PodcastDeleteFeed,
-    PodcastDeleteAllFeeds,
-    PodcastSearchAddFeed,
-    PodcastRefreshFeed,
-    PodcastRefreshAllFeeds,
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
