+++ tui/src/ui/components/config_editor/color.rs
@@ -26,7 +26,8 @@ use std::convert::From;
 use termusiclib::config::v2::tui::theme::styles::ColorTermusic;
 use termusiclib::config::v2::tui::theme::ThemeWrap;
 use termusiclib::config::SharedTuiSettings;
-use termusiclib::types::{ConfigEditorMsg, Id, IdConfigEditor, Msg};
+use termusiclib::ids::{Id, IdConfigEditor};
+use termusiclib::types::{ConfigEditorMsg, Msg};
 use tui_realm_stdlib::{Input, Label, Select, Table};
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
 use tuirealm::event::{Key, KeyEvent, KeyModifiers, NoUserEvent};
