+++ tui/src/ui/components/config_editor/general.rs
@@ -28,7 +28,7 @@ use crate::CombinedSettings;
 use anyhow::Result;
 use termusiclib::config::v2::tui::{keys::Keys, Alignment as XywhAlign};
 use termusiclib::config::SharedTuiSettings;
-use termusiclib::types::{Id, IdConfigEditor};
+use termusiclib::ids::{Id, IdConfigEditor};
 use tui_realm_stdlib::{Input, Radio};
 use tuirealm::props::{Alignment, BorderType, Borders, Color, InputType, Style};
 use tuirealm::{
