+++ tui/src/ui/components/config_editor/key_combo.rs
@@ -26,7 +26,8 @@ use std::fmt::Display;
 use anyhow::{bail, Result};
 use termusiclib::config::v2::tui::keys::{KeyBinding, Keys};
 use termusiclib::config::SharedTuiSettings;
-use termusiclib::types::{ConfigEditorMsg, Id, IdConfigEditor, IdKey, KFMsg, Msg};
+use termusiclib::ids::{Id, IdConfigEditor, IdKey};
+use termusiclib::types::{ConfigEditorMsg, KFMsg, Msg};
 use tui_realm_stdlib::utils::get_block;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
 use tuirealm::event::{Key, KeyEvent, KeyModifiers, NoUserEvent};
