+++ tui/src/ui/components/config_editor/mod.rs
@@ -32,7 +32,7 @@ use crate::ui::{ConfigEditorMsg, Msg};
 use anyhow::Result;
 
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
-use termusiclib::types::{Id, IdConfigEditor};
+use termusiclib::ids::{Id, IdConfigEditor};
 use tui_realm_stdlib::{Radio, Span};
 use tuirealm::props::{Alignment, BorderSides, BorderType, Borders, Style, TextSpan};
 use tuirealm::{event::NoUserEvent, Component, Event, MockComponent};
