+++ tui/src/ui/components/database.rs
@@ -1,9 +1,10 @@
 use std::path::Path;
 
 use termusiclib::config::SharedTuiSettings;
+use termusiclib::ids::Id;
 use termusiclib::library_db::const_unknown::{UNKNOWN_ARTIST, UNKNOWN_FILE, UNKNOWN_TITLE};
 use termusiclib::library_db::{Indexable, SearchCriteria, TrackDB};
-use termusiclib::types::{DBMsg, Id, Msg};
+use termusiclib::types::{DBMsg, Msg};
 use termusiclib::utils::{is_playlist, playlist_get_vec};
 use tui_realm_stdlib::List;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
