+++ tui/src/ui/components/podcast.rs
@@ -6,9 +6,10 @@ use sanitize_filename::{sanitize_with_options, Options};
 use serde_json::Value;
 use std::time::Duration;
 use termusiclib::config::SharedTuiSettings;
+use termusiclib::ids::Id;
 use termusiclib::podcast::{download_list, EpData, PodcastFeed, PodcastNoId};
 use termusiclib::track::MediaType;
-use termusiclib::types::{Id, Msg, PCMsg};
+use termusiclib::types::{Msg, PCMsg};
 use tokio::runtime::Handle;
 use tui_realm_stdlib::List;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
