+++ tui/src/ui/components/popups/general_search.rs
@@ -24,7 +24,8 @@
 use crate::ui::Model;
 use anyhow::{anyhow, bail, Result};
 use termusiclib::config::{SharedTuiSettings, TuiOverlay};
-use termusiclib::types::{GSMsg, Id, Msg};
+use termusiclib::ids::Id;
+use termusiclib::types::{GSMsg, Msg};
 use tui_realm_stdlib::{Input, Table};
 use tui_realm_treeview::TREE_INITIAL_NODE;
 use tuirealm::command::{Cmd, CmdResult, Direction, Position};
