+++ tui/src/ui/components/tag_editor/view.rs
@@ -30,8 +30,8 @@ use crate::ui::model::Model;
 use crate::ui::utils::{draw_area_in_absolute, draw_area_top_right_absolute};
 use anyhow::Result;
 use std::path::Path;
+use termusiclib::ids::{Id, IdTagEditor};
 use termusiclib::track::Track;
-use termusiclib::types::{Id, IdTagEditor};
 use tuirealm::props::{Alignment, AttrValue, Attribute, PropPayload, PropValue, TextSpan};
 use tuirealm::ratatui::layout::{Constraint, Direction, Layout};
 use tuirealm::ratatui::widgets::Clear;
