+++ tui/src/ui/model/mod.rs
@@ -9,6 +9,7 @@ use anyhow::{anyhow, bail, Context};
 use termusiclib::config::v2::tui::keys::Keys;
 use termusiclib::config::v2::tui::theme::ThemeWrap;
 use termusiclib::config::{ServerOverlay, SharedServerSettings, SharedTuiSettings};
+use termusiclib::ids::Id;
 use termusiclib::library_db::TrackDB;
 use termusiclib::library_db::{DataBase, SearchCriteria};
 use termusiclib::player::playlist_helpers::PlaylistTrackSource;
@@ -17,7 +18,7 @@ use termusiclib::podcast::{db::Database as DBPod, Podcast, PodcastFeed};
 use termusiclib::songtag::SongTag;
 use termusiclib::taskpool::TaskPool;
 use termusiclib::track::{MediaType, Track};
-use termusiclib::types::{Id, Msg, YoutubeOptions};
+use termusiclib::types::{Msg, YoutubeOptions};
 #[cfg(all(feature = "cover-ueberzug", not(target_os = "windows")))]
 use termusiclib::ueberzug::UeInstance;
 use termusiclib::utils::get_app_config_path;
