+++ tui/src/ui/model/update.rs
@@ -27,11 +27,12 @@ use crate::ui::{model::TermusicLayout, Model};
 use anyhow::anyhow;
 use std::path::Path;
 use std::time::Duration;
+use termusiclib::ids::{Id, IdTagEditor};
 use termusiclib::library_db::SearchCriteria;
 use termusiclib::track::MediaType;
 use termusiclib::types::{
-    DBMsg, DLMsg, GSMsg, Id, IdTagEditor, LIMsg, LyricMsg, MainLayoutMsg, Msg, PCMsg, PLMsg,
-    PlayerMsg, SavePlaylistMsg, XYWHMsg, YSMsg,
+    DBMsg, DLMsg, GSMsg, LIMsg, LyricMsg, MainLayoutMsg, Msg, PCMsg, PLMsg, PlayerMsg,
+    SavePlaylistMsg, XYWHMsg, YSMsg,
 };
 use tokio::runtime::Handle;
 use tokio::time::sleep;
