+++ tui/src/ui/model/youtube_options.rs
@@ -30,9 +30,10 @@ use std::path::{Path, PathBuf};
 use std::sync::{Arc, LazyLock};
 use std::thread;
 use std::time::Duration;
+use termusiclib::ids::Id;
 use termusiclib::invidious::Instance;
 use termusiclib::track::Track;
-use termusiclib::types::{DLMsg, Id, Msg};
+use termusiclib::types::{DLMsg, Msg};
 use termusiclib::types::{YSMsg, YoutubeOptions};
 use termusiclib::utils::get_parent_folder;
 use tokio::runtime::Handle;
